fix(trusty-agents-ui): split src-tauri main.rs under the 500-SLOC cap
Closes #3270
